% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_parks.R
\name{classify_parks}
\alias{classify_parks}
\title{Classify parks based on threshold}
\usage{
classify_parks(
  min_park_size = 10000,
  min_unmanaged_tree_area = 10000,
  min_vegetation_percentage = 80,
  v_parks_all = NA
)
}
\arguments{
\item{min_park_size}{Minimum park size to be filtered for, in m^2.}

\item{min_unmanaged_tree_area}{Minimum unmanaged trees area to be classified as secondary forest
habitat, in m^2.}

\item{min_vegetation_percentage}{Minimum percentage of vegetation for parks to be classified as
secondary forest habitat, in m^2.}
}
\value{
An sf object that contains the Singapore park geospatial layer.
}
\description{
\code{classify_parks} returns a sf object of parks with habitat column.
}
\details{
This function takes the Singapore park gpkg and reclassifies the geospatial files
based on rules defined by minimum park size \code{min_park_size}, minimum area for unmanaged
trees \code{min_unmanaged_tree_area}, minimum percentage of vegetation
\code{min_vegetation_percentage}.
}
\examples{
classify_parks(10000, 10000, 80, NA)
}
