% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_boolean_check_vars.R
\name{generate_boolean_check_vars}
\alias{generate_boolean_check_vars}
\title{Generate variables needed for boolean checks for each species}
\usage{
generate_boolean_check_vars(df_species, date_cut_off = as.Date("1960-01-01"))
}
\arguments{
\item{df_species}{A data.table containing individual species records.
Should minimally have specimen identifier, species identity, verbatim locality and/or
latitude and longitude, collection date and specimen type (worker, queen, male/ female alate).}

\item{date_cut_off}{The date cut-off for the bifurcation of the flow chart.
See Wang et al. (2020) Figure 1 for details.}
}
\value{
A data.table of variables used later in the boolean checks in the generate_iucn_status
function.
}
\description{
Generate variables needed for boolean checks for each species
}
\examples{
generate_boolean_check_vars(df_species, as.Date("1960-01-01"))
}
