% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_habitat.R
\name{generate_habitat}
\alias{generate_habitat}
\title{Generate the site name and corresponding habitat category for each species record/row}
\usage{
generate_habitat(
  df_species,
  identifier_columns = c("id", "species", "type"),
  coord_columns = c("X", "Y"),
  df_species_epsg = 4236,
  vector_layers = list(v_islands = NA, v_parks_nat_res = NA, v_parks_all = NA,
    v_greenery = NA, v_planning_areas = NA)
)
}
\arguments{
\item{df_species}{A data.table containing individual species records.
Should minimally have specimen identifier, species identity, verbatim locality and/or
latitude and longitude, collection date and specimen type (worker, queen, male/ female alate).}

\item{identifier_columns}{These are a group of identifier columns, used for easy reference
and joins to other derived datasets. Should include columns analgous to "id", "species",
"type" (i.e. whether alate, worker etc).}

\item{coord_columns}{The columns names of the "X" and "Y" coordinates respectively.}

\item{df_species_epsg}{The EPSG number of the coordinates. This would typically be
4326 for latitude and longitude, representing WGS84.}

\item{vector_layers}{List of sf objects, containing islands (v_islands), parks \link{nature reserves}
(v_parks_nat_res), parks \link{all} (v_parks_all), unmanaged greenery (v_greenery), URA planning areas
(v_planning_areas).
If none are specified (i.e. they are NA), the function will used a preset group of vectors
which are loaded from the package itself through \code{data(vector_name)}.}
}
\value{
A data.table with appended site name and habitat type it belongs to.
}
\description{
Generate the site name and corresponding habitat category (habitat_final) based on a series of
spatial joins from point data (typically latitude and longitude, though this can be changed) to
the geospatial vectors found in the package.
}
