% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_habitat_sp_matrix.R
\name{generate_habitat_sp_matrix}
\alias{generate_habitat_sp_matrix}
\title{Generate habitat versus species matrix}
\usage{
generate_habitat_sp_matrix(
  df_habitat,
  collection_date_column = "collection_date",
  date_cut_off = as.Date("1960-01-01")
)
}
\arguments{
\item{df_habitat}{The data.table with records of ant species, site name and its corresponding
habitat type.}

\item{collection_date_column}{The name of the collection date column.}

\item{date_cut_off}{The cut-off (maximum) threshold date for the collection date (inclusive).}
}
\value{
A data.table that contains species (rows) versus habitat (columns) with values
representing number of unique sites it is found in for that habitat type.
}
\description{
Generates the habitat versus species matrix, by counting the number of unique sites,
based on the record-habitat matrix with the site name and corresponding habitat type columns.
}
\examples{
generate_habitat_matrix(df_habitat, "collection_date", as.Date("1960-01-01"))
}
