% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_iucn_status.R
\name{generate_iucn_status}
\alias{generate_iucn_status}
\title{Generate the IUCN status for each species}
\usage{
generate_iucn_status(df_bool, df_habitat_sp_mat, site_cut_off = 2)
}
\arguments{
\item{df_bool}{data.table output from generate_boolean_check_vars function. It is a data.table
of variables that will be used in boolean checks within this function.}

\item{df_habitat_sp_mat}{data.table output from generate_habitat_sp_matrix function. It is a
data.table that contains the}

\item{site_cut_off}{The cut-off (maximum) threshold number of sites.}
}
\value{
A data.table with species and their corresponding boolean checks and assigned
IUCN status.
}
\description{
Generates the IUCN status for each species based on the data.table of species-habitat matrix
from generate_habitat_sp_matrix function and the data.table of variables associated with
boolean checks from generate_boolean_check_vars function.
}
