% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_site_names.R
\name{generate_site_names}
\alias{generate_site_names}
\title{Generate site names from geospatial layers}
\usage{
generate_site_names()
}
\value{
data.table of park names
}
\description{
Generate site names from geospatial layers
}
