% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sg_ants_test}
\alias{sg_ants_test}
\title{sg_ants_test}
\format{
An data.table with 313 rows and 6 columns:
\describe{
\item{species}{species name}
\item{X}{X coordinate (i.e. longitude)}
\item{Y}{Y coordinate (i.e. latitude)}
\item{collection_date}{collection date in YYYY-MM-DD}
\item{type}{type of specimen (worker, reproductive)}
...
}
}
\source{
\url{https://github.com/eunices/sgAnts/}
}
\usage{
sg_ants_test
}
\description{
A dataset containing simulated ant distribution data in Singapore
}
\keyword{datasets}
