% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v_islands}
\alias{v_islands}
\title{v_islands}
\format{
An sf object with 43 rows and 3 columns:
\describe{
\item{NAME}{Site name}
\item{geom}{Points within sf object}
...
}
}
\source{
\url{https://github.com/eunices/sgAnts/}
}
\usage{
v_islands
}
\description{
A geospatial vector layer for islands in Singapore from the basemap of Singapore
by Teo Siyang.
}
\keyword{datasets}
