% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v_parks_all}
\alias{v_parks_all}
\title{v_parks_all}
\format{
An sf object with 557 rows and 17 columns:
\describe{
\item{NAME}{Site name}
\item{geom}{Points within sf object}
\item{habitat}{Assigned habitat based on rules outlined in Wang et al.}
\item{AREA_SQ_M}{Area of site, in sq km.}
\item{prop_veg_canopy_unmanaged}{Proportion of pixels of unmanaged vegetation with canopy
of entire site. }
\item{prop_veg_canopy}{Proportion of pixels of vegetation with canopy of entire site.}
\item{prop_veg}{Proportion of pixels with vegetation of entire site.}
\item{prop_veg_no_canopy_of_veg}{Proportion of vegetation with no canopy based on counting
pixels from Gaw et al. (2019)}
\item{mangrove}{Number of pixels which is mangrove based on Gaw et al. (2019)}
\item{veg_canopy_unmanaged}{Number of pixels which is unmanaged greenery with canopy based on
Gaw et al. (2019)}
...
}
}
\source{
\url{https://data.gov.sg/dataset/sdcp-park?resource_id=9584bbb8-bd5c-49fe-ac58-55734fb56538}
}
\usage{
v_parks_all
}
\description{
A geospatial vector layer for parks in Singapore from SDCP parks accessed on May 2020  from
https://data.gov.sg/dataset/sdcp-park?resource_id=9584bbb8-bd5c-49fe-ac58-55734fb56538
}
\keyword{datasets}
