% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v_parks_nat_res}
\alias{v_parks_nat_res}
\title{v_parks_nat_res}
\format{
An sf object with 9 rows and 13 columns:
\describe{
\item{NAME}{Site name}
\item{geom}{Points within sf object}
...
}
}
\source{
\url{https://github.com/eunices/sgAnts/}
}
\usage{
v_parks_nat_res
}
\description{
A geospatial vector layer for parks (nature reserves) in Singapore from the basemap of Singapore
by Teo Siyang.
}
\keyword{datasets}
