% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v_planning_areas}
\alias{v_planning_areas}
\title{v_planning_areas}
\format{
An sf object with 52 rows and 14 columns:
\describe{
\item{NAME}{Site name}
\item{geom}{Points within sf object}
...
}
}
\source{
\url{https://data.gov.sg/dataset/master-plan-2014-planning-area-boundary-no-sea/}
}
\usage{
v_planning_areas
}
\description{
A geospatial vector layer for planning areas in Singapore from
https://data.gov.sg/dataset/master-plan-2014-planning-area-boundary-no-sea.
}
\keyword{datasets}
