% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v_greenery}
\alias{v_greenery}
\title{v_greenery}
\format{
An sf object with 449 rows and 3 columns:
\describe{
\item{NAME}{Site name}
\item{geom}{Points within sf object}
\item{AREA_SQ_M}{Area of greenery patch, in sq km}
...
}
}
\source{
\url{https://www.mdpi.com/2306-5729/4/3/116/}
}
\usage{
v_greenery
}
\description{
A geospatial vector layer for unmanaged greenery (with tree cover) patches in Singapore based on
Gaw et al. (2019) https://www.mdpi.com/2306-5729/4/3/116.
}
\keyword{datasets}
