% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{v_parks_all}
\alias{v_parks_all}
\title{v_parks_all}
\format{
An sf object with 557 rows and 17 columns:
\describe{
\item{NAME}{Site name}
\item{geom}{Points within sf object}
...
}
}
\source{
\url{https://data.gov.sg/dataset/sdcp-park?resource_id=9584bbb8-bd5c-49fe-ac58-55734fb56538}
}
\usage{
v_parks_all
}
\description{
A geospatial vector layer for parks in Singapore from SDCP parks accessed on May 2020  from
https://data.gov.sg/dataset/sdcp-park?resource_id=9584bbb8-bd5c-49fe-ac58-55734fb56538
}
\keyword{datasets}
